removed socket portion from requests_stdlib and moved it into file
